+++ src/core/input_handler.mjs
@@ -1,4 +1,7 @@
+import { Ticker, UPDATE_PRIORITY } from "pixi.js";
+
 const keys = new Set();
+const pressedKeys = new Set();
 const mouse = {
 	x: 0,
 	y: 0,
@@ -7,6 +10,15 @@ const mouse = {
 	pressed: 0
 };
 
+const ticker = new Ticker();
+ticker.maxFPS = 60;
+ticker.autoStart = true;
+ticker.add(resetPress, null, UPDATE_PRIORITY.LOW);
+
+function resetPress(ticker) {
+	prevTickKeys.clear();
+}
+
 addEventListener("keydown", keydown);
 addEventListener("keyup", keyup);
 addEventListener("mousemove", mousemove);
@@ -14,8 +26,10 @@ addEventListener("mousedown", mousedown);
 addEventListener("mouseup", mousedown); // they do the same thing!!!
 
 function keydown(e) {
-	if (!e.repeat)
+	if (!e.repeat) {
 		keys.add(e.code);
+		ticker.addOnce((ticker) => { prevTickKeys.add(e.code); }, null, UPDATE_PRIORITY.HIGH);
+	}
 }
 
 function keyup(e) {
@@ -33,4 +47,4 @@ function mousemove(e) {
 	mouse.moveY = e.movementY;
 }
 
-export { keys, mouse };
+export { keys, pressedKeys, mouse };