+++ changelog.md
@@ -9,6 +9,7 @@
 - Rewrote asset loading system
 - Rewrote text system
 	- For reasons beyond me the health and score no longer align to the pixel grid
+- Updated input handling to process single presses, not just holds
 
 ## v0.02 (25/12/31)